Flower Classification/requirements.txt
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.2.2
Pillow==10.0.0
numpy==1.25.0